Updated Controller for Resources
